move trace function into tracing.py
